+++ src/index.ts
@@ -1,9 +1,26 @@
-import { isDefined } from './guards'
+import {
+  isDefined,
+  isString,
+  isNumber,
+  isBigInt,
+  isBool,
+  isSymbol,
+  isNull,
+} from './guards'
 import { generateError } from './utils'
 
 import type { JSONValue } from './JSON'
 import type { Tuple } from './tuple'
 import type { PropsOf } from './vue'
 
-export { isDefined, generateError }
+export {
+  isDefined,
+  isString,
+  isNumber,
+  isBigInt,
+  isBool,
+  isSymbol,
+  isNull,
+  generateError,
+}
 export type { JSONValue, Tuple, PropsOf }